biggest commit ever
literally reformatting the whole app
