+++ app/models/broadcast.js
@@ -1,0 +1,25 @@
+var mongoose = require('mongoose'),
+    shortid = require('shortid'),
+    helpers = require('./model_helpers'),
+    Schema = mongoose.Schema;
+
+var broadcast = new Schema({
+	// broadcast about the object with given url
+  _id: {
+    type: String,
+    default: shortid.generate
+  },
+  url: {
+    type: String,
+		required: true
+  },
+  message: {
+    type: String,
+    required: true
+  }
+}, { collection: 'broadcasts', timestamps: true });
+
+broadcast.statics.find =
+  helpers.finderForProperty("_id", { findOne: true, caseInsensitive: false });
+
+module.exports = mongoose.model('Broadcast', broadcast);
