+++ app/models/model_helpers.js
@@ -1,0 +1,21 @@
+var mongoose = require('mongoose');
+
+module.exports.finderForProperty = function(field, props) {
+    if (props === undefined) {
+        props = {};
+    }
+
+    var isFindOne = props['findOne'] === true;
+    var caseInsensitive = props['caseInsensitive'] === true;
+
+    return function(value, cb) {
+        var lookupHash = {};
+        lookupHash[field] = caseInsensitive ? new RegExp('^' + value + '$', 'i') : value;
+
+        if(isFindOne) {
+            this.findOne(lookupHash, cb);
+        } else {
+            this.find(lookupHash, cb);
+        }
+    };
+};
