+++ package.json
@@ -12,6 +12,7 @@
     "body-parser": "^1.15.2",
     "chai": "^3.5.0",
     "connect-flash": "*",
+    "cookie-parser": "^1.4.3",
     "express": "^4.14.0",
     "express-handlebars": "*",
     "express-messages": "*",
